pyproject.toml
[tool.poetry]
name = "reprobench"
version = "0.6.0.dev7"
readme = "README.md"
description = "Reproducible Benchmark for Everyone"
authors = ["Rakha Kanz Kautsar <rkkautsar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6"
peewee = "^3.9"
strictyaml = "^1.0"
pynisher = "^0.5.0"
loguru = "^0.2.5"
psmon = "^1.1.0"
click = "^7.0"
apsw = "^3.9"
tqdm = "^4.31"
pathspec = "^0.5.9"
requests = "^2.21"
psutil = "^5.6"
py-cpuinfo = ">=4,<6"
msgpack-python = "^0.5.6"
pyzmq = "^18.0"
gevent = "^1.4"

[tool.poetry.dev-dependencies]
black = "=19.3b0"

[tool.poetry.scripts]
reprobench = "reprobench.console.main:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
